+++ Arbiter.App/Models/DebugSettings.cs
@@ -6,6 +6,7 @@ public class DebugSettings : ICloneable
 {
     public bool ShowDialogId { get; set; }
     public bool ShowPursuitId { get; set; }
+    public bool ShowDialogItemQuantity { get; set; }
     public bool ShowNpcId { get; set; }
     public bool ShowMonsterId { get; set; }
     public bool ShowMonsterClickId { get; set; }
@@ -24,6 +25,7 @@ public class DebugSettings : ICloneable
     {
         ShowDialogId = ShowDialogId,
         ShowPursuitId = ShowPursuitId,
+        ShowDialogItemQuantity = ShowDialogItemQuantity,
         ShowNpcId = ShowNpcId,
         ShowMonsterId = ShowMonsterId,
         ShowMonsterClickId = ShowMonsterClickId,