+++ Arbiter.App/Models/InventoryItem.cs
@@ -4,9 +4,10 @@ namespace Arbiter.App.Models;
 
 public readonly struct InventoryItem
 {
-    public static InventoryItem Empty => new() { IsEmpty = true, Name = string.Empty };
+    public static InventoryItem Empty(int slot = 0) => new() { Slot = slot, IsEmpty = true, Name = string.Empty };
 
     public bool IsEmpty { get; private init; }
+    public int Slot { get; init; }
     public ushort Sprite { get; init; }
     public byte Color { get; init; }
     public required string Name { get; init; }