+++ Arbiter.App/ViewModels/SettingsViewModel.cs
@@ -41,6 +41,7 @@ public partial class SettingsViewModel : ViewModelBase, IDialogResult<ArbiterSet
     [NotifyPropertyChangedFor(nameof(TraceMaxHistory))]
     [NotifyPropertyChangedFor(nameof(DebugShowDialogId))]
     [NotifyPropertyChangedFor(nameof(DebugShowPursuitId))]
+    [NotifyPropertyChangedFor(nameof(DebugShowDialogItemQuantity))]
     [NotifyPropertyChangedFor(nameof(DebugShowNpcId))]
     [NotifyPropertyChangedFor(nameof(DebugShowMonsterId))]
     [NotifyPropertyChangedFor(nameof(DebugShowMonsterClickId))]
@@ -196,6 +197,17 @@ public partial class SettingsViewModel : ViewModelBase, IDialogResult<ArbiterSet
         set
         {
             Settings.Debug.ShowPursuitId = value;
+            OnPropertyChanged();
+            HasChanges = true;
+        }
+    }
+
+    public bool DebugShowDialogItemQuantity
+    {
+        get => Settings.Debug.ShowDialogItemQuantity;
+        set
+        {
+            Settings.Debug.ShowDialogItemQuantity = value;
             OnPropertyChanged();
             HasChanges = true;
         }