+++ Arbiter.Net/Proxy/ProxyConnection.cs
@@ -4,7 +4,6 @@ using System.Threading.Channels;
 using Arbiter.Net.Client;
 using Arbiter.Net.Client.Messages;
 using Arbiter.Net.Filters;
-using Arbiter.Net.Serialization;
 using Arbiter.Net.Server;
 using Arbiter.Net.Server.Messages;
 
@@ -74,85 +73,6 @@ public partial class ProxyConnection : IDisposable
 
         _clientMessageFactory = clientMessageFactory ?? ClientMessageFactory.Default;
         _serverMessageFactory = serverMessageFactory ?? ServerMessageFactory.Default;
-    }
-
-    public bool EnqueuePacket(NetworkPacket packet, NetworkPriority priority = NetworkPriority.Normal)
-    {
-        // Prioritize outgoing client heartbeat packets
-        if (packet is ClientPacket { Command: ClientCommand.Heartbeat })
-        {
-            priority = NetworkPriority.High;
-        }
-
-        var writer = priority switch
-        {
-            NetworkPriority.High => _prioritySendQueue.Writer,
-            _ => _sendQueue.Writer,
-        };
-
-        if (!writer.TryWrite(packet))
-        {
-            return false;
-        }
-
-        PacketQueued?.Invoke(this, new NetworkPacketEventArgs(packet));
-        return true;
-    }
-
-    public bool EnqueueMessage(IClientMessage message, NetworkPriority priority = NetworkPriority.Normal)
-    {
-        var command = _clientMessageFactory.GetMessageCommand(message.GetType());
-        if (command is null)
-        {
-            return false;
-        }
-
-        var builder = new NetworkPacketBuilder(command.Value);
-        try
-        {
-            message.Serialize(ref builder);
-            var packet = builder.ToPacket();
-
-            // Prioritize outgoing client heartbeat packets
-            if (packet is ClientPacket { Command: ClientCommand.Heartbeat })
-            {
-                priority = NetworkPriority.High;
-            }
-
-            return EnqueuePacket(packet, priority);
-        }
-        finally
-        {
-            builder.Dispose();
-        }
-    }
-
-    public bool EnqueueMessage(IServerMessage message, NetworkPriority priority = NetworkPriority.Normal)
-    {
-        var command = _serverMessageFactory.GetMessageCommand(message.GetType());
-        if (command is null)
-        {
-            return false;
-        }
-
-        var builder = new NetworkPacketBuilder(command.Value);
-        try
-        {
-            message.Serialize(ref builder);
-            var packet = builder.ToPacket();
-
-            // Prioritize incoming server heartbeat packets
-            if (packet is ServerPacket { Command: ServerCommand.Heartbeat })
-            {
-                priority = NetworkPriority.High;
-            }
-
-            return EnqueuePacket(packet, priority);
-        }
-        finally
-        {
-            builder.Dispose();
-        }
     }
 
     internal async Task ConnectToRemoteAsync(IPEndPoint remoteEndpoint, CancellationToken token = default)